chore: setup the project
This commit setup the project using hardhat. The contract, his associated
test and his associated deploy script are the default ones created by hardhat.
Eslint, prettier and solhint quality librairies are configured and exposed
via npm scripts. Hardhat is configured to load environnement variables from
.env files and use them to interact with the local and ropsten networks,
generate a gas report and interact with the etherscan API.
Chai, Mocha and ethers are the librairies used to write the tests.
To conclude, the project is configured to allow its execution locally
by exposing a JSON-RPC.